Move spec-build to scripts
