+++ scripts/build_spec.js
@@ -4,9 +4,14 @@ const esbuild = require('esbuild')
 const fs = require('fs')
 const path = require('path')
 
-const entries = fs.readdirSync('spec')
+const entries = fs.readdirSync('spec/src')
 	.filter(dir => dir.endsWith('.spec.ts'))
-	.map(dir => path.join('spec', dir))
+	.map(dir => path.join('spec/src', dir))
+
+console.log(`------------------------------------
+SPECIFICATION USES LAST BUILD
+BE SURE TO BUILD BEFORE RUNNING SPEC
+------------------------------------`)
 
 esbuild.buildSync({
 	entryPoints: entries,
@@ -15,5 +20,5 @@ esbuild.buildSync({
 	minify: true,
 	platform: 'node',
 	write: true,
-	outdir: 'spec/_build'
+	outdir: 'spec/build'
 })
