+++ tsconfig.json
@@ -1,7 +1,7 @@
 {
 	"compilerOptions": {
 		"module": "commonjs",
-		"target": "ES6",
+		"target": "ESNext",
 		"outDir": "build",
 		"lib": ["ES6", "DOM"],
 		"sourceMap": true,
